test: relocate snapshot harness to TestUtils, pin scale via traits, shrink hero ref, harden record guard
