+++ Alfie/AlfieKit/Package.swift
@@ -310,7 +310,6 @@ let package = Package(
             dependencies: [
                 .product(name: "CombineSchedulers", package: "combine-schedulers"),
                 .product(name: "SnapshotTesting", package: "swift-snapshot-testing"),
-                "Utils",
             ]
         ),
