+++ Alfie/AlfieKit/Sources/Home/Resources/HomeAssets.xcassets/hero-placeholder.imageset/Contents.json
@@ -1,0 +1,6 @@
+{
+  "images" : [
+    { "filename" : "hero-placeholder.png", "idiom" : "universal" }
+  ],
+  "info" : { "author" : "xcode", "version" : 1 }
+}
