+++ Alfie/AlfieKit/Sources/TestUtils/Helpers/Snapshotting+Extension.swift
@@ -8,6 +8,15 @@ extension Snapshotting where Value == UIView, Format == UIImage {
         precision: Float = 0.9,
         perceptualPrecision: Float = 0.95
     ) -> Snapshotting {
-        .image(precision: precision, perceptualPrecision: perceptualPrecision, traits: .init(displayGamut: .SRGB))
+        // Pin displayScale to 3 in the strategy's traits. The renderer would otherwise fall back to
+        // UIScreen.main.scale, so pinning here keeps references comparable without mutating global state.
+        .image(
+            precision: precision,
+            perceptualPrecision: perceptualPrecision,
+            traits: UITraitCollection(traitsFrom: [
+                UITraitCollection(displayScale: 3),
+                UITraitCollection(displayGamut: .SRGB),
+            ])
+        )
     }
 }
