+++ Alfie/AlfieKit/Tests/HomeTests/HomeViewSnapshotTests.swift
@@ -21,6 +21,11 @@ final class HomeViewSnapshotTests: XCTestCase {
     }
 
     func test_homeView_withHeroBanners() {
+        // Use a flat-colour placeholder asset instead of the photographic hero JPGs: the test covers
+        // carousel layout, and a solid image keeps the reference PNG small and its diffs meaningful.
+        mockViewModel.heroBanners = HomeHeroBanner.placeholders.map {
+            HomeHeroBanner(id: $0.id, imageName: "hero-placeholder", title: $0.title)
+        }
         let sut = HomeView(viewModel: mockViewModel)
         assertSnapshot(of: sut.embededInContainer(),
                        as: .defaultImage(),
