+++ Docs/Testing.md
@@ -55,4 +55,4 @@ final class FeatureServiceTests: XCTestCase {
 - Tests live in the AlfieKit module test targets; shared helpers are in `TestUtils`
 - References are pinned to iOS major 26 and run as part of `./Alfie/scripts/verify.sh`
 
-See `Docs/SnapshotTesting.md` for the device/precision policy, the record loop, and the migration playbook.
+See `Docs/SnapshotTesting.md` for the device/precision policy, the record loop, and where the tests live.
